refactor(api): extract HTTP transport helpers from client.rs into transport.rs
`client.rs` had grown to hold three concerns: the API client itself, the
retry / backoff state machine, and the response-bounding + parse-dispatch
plumbing. The latter two are pure transport — they don't reference the
client's state and exist only to harden every HTTP round-trip.

Lift them verbatim into `src/api/transport.rs`:

  * `RETRY_MAX_ATTEMPTS`, `RETRY_BASE_DELAY_MS`
  * `is_retryable_status`, `is_retryable_error`
  * `backoff_delay`, `retry_after_secs`
  * `MAX_RESPONSE_BYTES`, `PARSE_BLOCKING_THRESHOLD`
  * `parse_json_maybe_blocking`, `read_bounded_body`

All exposed at `pub(super)` so only the `api` module sees them; no public
surface changes. `client.rs` keeps a single `use super::transport::{...}`
and the entire bodies of those helpers — including their audit-trail doc
comments — survive the move unmodified.

Zero behavioural change: same bytes through the same code paths, same
retry policy, same 32 MiB body cap, same 256 KiB serde-blocking threshold.
Verified by `cargo test --all-targets` (619 lib + ~440 integration green;
the http_429 / http_502 / http_503 retry tests in `error_handling_e2e`
exercise the lifted code end-to-end).
